got subsys_server executable to compile but there are some WEIRD things going on in the code with regards to mixing C and C++
