+++ subsys_server_NEW/CMakeLists.txt
@@ -142,6 +142,9 @@ endforeach(lang ${${EXE}_languages})
 # LINK AGAINST THE NECESSARY EXERARIES 
 ################################################################################
 target_link_libraries(${EXE} PRIVATE CLIENT_API)
+target_link_libraries(${EXE} PRIVATE LIB_IPC)
+
+
 
 find_package(Threads REQUIRED)
 if(Threads_FOUND)
